Code cleanup

+++ src/main/java/gregtech/api/objects/GT_RenderedTexture.java
@@ -71,10 +71,10 @@ public class GT_RenderedTexture implements ITexture, IColorModulationContainer {
         IIcon aIcon = mIconContainer.getIcon();
 
         float d_16 = 16.0F;
-        float d3 = (float)aIcon.getInterpolatedU(aRenderer.renderMaxX * d_16);
-        float d4 = (float)aIcon.getInterpolatedU(aRenderer.renderMinX * d_16);
-        float d5 = (float)aIcon.getInterpolatedV(aRenderer.renderMinZ * d_16);
-        float d6 = (float)aIcon.getInterpolatedV(aRenderer.renderMaxZ * d_16);
+        float d3 = aIcon.getInterpolatedU(aRenderer.renderMaxX * d_16);
+        float d4 = aIcon.getInterpolatedU(aRenderer.renderMinX * d_16);
+        float d5 = aIcon.getInterpolatedV(aRenderer.renderMinZ * d_16);
+        float d6 = aIcon.getInterpolatedV(aRenderer.renderMaxZ * d_16);
 
         if (aRenderer.renderMinX < 0.0D || aRenderer.renderMaxX > 1.0D) {
             d3 = aIcon.getMaxU();
@@ -92,18 +92,18 @@ public class GT_RenderedTexture implements ITexture, IColorModulationContainer {
         float d14 = aZ + (float)aRenderer.renderMinZ;
         float d15 = aZ + (float)aRenderer.renderMaxZ;
 
-        Tessellator.instance.addVertexWithUV((double)d11, (double)d13, (double)d15, (double)d3, (double)d6);
-        Tessellator.instance.addVertexWithUV((double)d11, (double)d13, (double)d14, (double)d3, (double)d5);
-        Tessellator.instance.addVertexWithUV((double)d12, (double)d13, (double)d14, (double)d4, (double)d5);
-        Tessellator.instance.addVertexWithUV((double)d12, (double)d13, (double)d15, (double)d4, (double)d6);
+        Tessellator.instance.addVertexWithUV(d11, d13, d15, d3, d6);
+        Tessellator.instance.addVertexWithUV(d11, d13, d14, d3, d5);
+        Tessellator.instance.addVertexWithUV(d12, d13, d14, d4, d5);
+        Tessellator.instance.addVertexWithUV(d12, d13, d15, d4, d6);
 
         if ((aIcon = mIconContainer.getOverlayIcon()) != null) {
             Tessellator.instance.setColorRGBA(128, 128, 128, 255);
 
-            Tessellator.instance.addVertexWithUV((double)d11, (double)d13, (double)d15, (double)d3, (double)d6);
-            Tessellator.instance.addVertexWithUV((double)d11, (double)d13, (double)d14, (double)d3, (double)d5);
-            Tessellator.instance.addVertexWithUV((double)d12, (double)d13, (double)d14, (double)d4, (double)d5);
-            Tessellator.instance.addVertexWithUV((double)d12, (double)d13, (double)d15, (double)d4, (double)d6);
+            Tessellator.instance.addVertexWithUV(d11, d13, d15, d3, d6);
+            Tessellator.instance.addVertexWithUV(d11, d13, d14, d3, d5);
+            Tessellator.instance.addVertexWithUV(d12, d13, d14, d4, d5);
+            Tessellator.instance.addVertexWithUV(d12, d13, d15, d4, d6);
         }
     }
